+++ services/workflows-service/src/common/filters/prisma-client-validation-filter/utils/parsers/unknown-argument-parser/unknown-argument.parser.unit.test.ts
@@ -41,6 +41,7 @@ describe('UnknownArgumentParser', () => {
             params.failedAtPath,
             params.type,
           );
+
           return result;
         }, {} as Record<string, string>);
       });