+++ services/workflows-service/src/providers/file/file.service.ts
@@ -55,6 +55,7 @@ export class FileService {
           mimeType,
         };
       }
+
       const filePaths = await this.copyThroughFileSystem(
         sourceServiceProvider,
         sourceRemoteFileConfig,