+++ services/workflows-service/src/storage/storage.controller.internal.ts
@@ -4,7 +4,7 @@ import { FileInterceptor } from '@nestjs/platform-express';
 import * as swagger from '@nestjs/swagger';
 import { ApiBody, ApiConsumes } from '@nestjs/swagger';
 import type { Response } from 'express';
-import * as nestAccessControl from 'nest-access-control';
+// import * as nestAccessControl from 'nest-access-control';
 import { StorageService } from './storage.service';
 import * as errors from '../errors';
 import { fileFilter } from './file-filter';
@@ -35,8 +35,8 @@ import { getFileMetadata } from '@/common/get-file-metadata/get-file-metadata';
 export class StorageControllerInternal {
   constructor(
     protected readonly service: StorageService,
-    @nestAccessControl.InjectRolesBuilder()
-    protected readonly rolesBuilder: nestAccessControl.RolesBuilder,
+    // @nestAccessControl.InjectRolesBuilder()
+    // protected readonly rolesBuilder: nestAccessControl.RolesBuilder,
     protected readonly logger: AppLoggerService,
     protected readonly httpService: HttpService,
   ) {}
@@ -132,6 +132,7 @@ export class StorageControllerInternal {
         fileNameInBucket: persistedFile.fileNameInBucket,
         mimeType,
       });
+
       return res.json({ signedUrl, mimeType });
     }
 
@@ -159,6 +160,7 @@ export class StorageControllerInternal {
     if (!path.isAbsolute(filePath)) return filePath;
 
     const rootDir = path.parse(os.homedir()).root;
+
     return path.join(rootDir, filePath);
   }