+++ services/workflows-service/src/workflow/dtos/workflow-definition-update-input.ts
@@ -51,8 +51,4 @@ export class WorkflowDefinitionUpdateInput {
   @IsString()
   @IsOptional()
   assigneeId?: string;
-
-  @IsString()
-  @IsOptional()
-  postUpdateEventName?: string;
 }