+++ services/workflows-service/src/workflow/types/params.ts
@@ -1,12 +1,4 @@
-import type { TProjectIds } from '@/types';
-
 export interface FindLastActiveFlowParams {
   workflowDefinitionId: string;
   businessId: string;
-}
-
-export interface GetLastActiveFlowParams {
-  email: string;
-  workflowRuntimeDefinitionId: string;
-  projectIds: TProjectIds;
 }