+++ apps/kyb-app/CHANGELOG.md
@@ -1,5 +1,48 @@
 # kyb-app
 
+## 0.1.49
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.7.48
+  - @ballerine/workflow-browser-sdk@0.5.48
+
+## 0.1.48
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.7.47
+  - @ballerine/workflow-browser-sdk@0.5.47
+
+## 0.1.47
+
+### Patch Changes
+
+- version bump
+- Updated dependencies
+  - @ballerine/common@0.7.46
+  - @ballerine/blocks@0.1.28
+  - @ballerine/ui@0.3.30
+  - @ballerine/workflow-browser-sdk@0.5.46
+
+## 0.1.46
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.7.45
+  - @ballerine/workflow-browser-sdk@0.5.45
+
+## 0.1.45
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.7.44
+  - @ballerine/workflow-browser-sdk@0.5.44
+
 ## 0.1.44
 
 ### Patch Changes