+++ services/workflows-service/CHANGELOG.md
@@ -1,5 +1,51 @@
 # @ballerine/workflows-service
 
+## 0.5.48
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.7.48
+  - @ballerine/workflow-core@0.5.48
+  - @ballerine/workflow-node-sdk@0.5.48
+
+## 0.5.47
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/workflow-core@0.5.47
+  - @ballerine/common@0.7.47
+  - @ballerine/workflow-node-sdk@0.5.47
+
+## 0.5.46
+
+### Patch Changes
+
+- version bump
+- Updated dependencies
+  - @ballerine/common@0.7.46
+  - @ballerine/workflow-core@0.5.46
+  - @ballerine/workflow-node-sdk@0.5.46
+
+## 0.5.45
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.7.45
+  - @ballerine/workflow-core@0.5.45
+  - @ballerine/workflow-node-sdk@0.5.45
+
+## 0.5.44
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/workflow-core@0.5.44
+  - @ballerine/common@0.7.44
+  - @ballerine/workflow-node-sdk@0.5.44
+
 ## 0.5.43
 
 ### Patch Changes